JE-403: Started work on better exception handling. Cleaned up duplicated code issues. Moved sync code to DataProviderService
