+++ src/Command/SyncAccountsCommand.php
@@ -3,6 +3,7 @@
 namespace App\Command;
 
 use App\Service\BillingService;
+use App\Service\DataProviderService;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -15,7 +16,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 )]
 class SyncAccountsCommand extends Command
 {
-    public function __construct(private readonly BillingService $billingService)
+    public function __construct(private readonly DataProviderService $dataProviderService)
     {
         parent::__construct($this->getName());
     }
@@ -28,7 +29,7 @@ class SyncAccountsCommand extends Command
     {
         $io = new SymfonyStyle($input, $output);
 
-        $this->billingService->syncAccounts(function ($i, $length) use ($io) {
+        $this->dataProviderService->syncAccounts(function ($i, $length) use ($io) {
             if (0 == $i) {
                 $io->progressStart($length);
             } elseif ($i >= $length - 1) {
