+++ src/Command/SyncIssuesCommand.php
@@ -4,6 +4,7 @@ namespace App\Command;
 
 use App\Repository\ProjectRepository;
 use App\Service\BillingService;
+use App\Service\DataProviderService;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -17,7 +18,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 class SyncIssuesCommand extends Command
 {
     public function __construct(
-        private readonly BillingService $billingService,
+        private readonly DataProviderService $dataProviderService,
         private readonly ProjectRepository $projectRepository,
     ) {
         parent::__construct($this->getName());
@@ -43,7 +44,7 @@ class SyncIssuesCommand extends Command
         foreach ($projects as $project) {
             $io->writeln("Processing issues for {$project->getName()}");
 
-            $this->billingService->syncIssuesForProject($project->getId(), function ($i, $length) use ($io) {
+            $this->dataProviderService->syncIssuesForProject($project->getId(), function ($i, $length) use ($io) {
                 if (0 == $i) {
                     $io->progressStart($length);
                 } elseif ($i >= $length - 1) {
