+++ src/Controller/HomeController.php
@@ -11,8 +11,8 @@ use Symfony\Component\Routing\Annotation\Route;
 class HomeController extends AbstractController
 {
     #[Route('/', name: 'index')]
-    public function index(Request $request): Response
+    public function index(): Response
     {
-        return $this->render('home/index.html.twig', []);
+        return $this->render('home/index.html.twig');
     }
 }
