+++ src/Service/BillingService.php
@@ -2,231 +2,36 @@
 
 namespace App\Service;
 
-use App\Entity\Account;
-use App\Entity\Client;
 use App\Entity\Invoice;
 use App\Entity\InvoiceEntry;
-use App\Entity\Issue;
-use App\Entity\Project;
-use App\Entity\Version;
-use App\Entity\Worklog;
 use App\Enum\ClientTypeEnum;
 use App\Enum\InvoiceEntryTypeEnum;
 use App\Exception\InvoiceAlreadyOnRecordException;
-use App\Repository\AccountRepository;
-use App\Repository\ClientRepository;
+use App\Exception\EconomicsException;
 use App\Repository\InvoiceEntryRepository;
 use App\Repository\InvoiceRepository;
-use App\Repository\IssueRepository;
-use App\Repository\ProjectRepository;
-use App\Repository\VersionRepository;
-use App\Repository\WorklogRepository;
-use Doctrine\ORM\EntityManagerInterface;
+use PhpOffice\PhpSpreadsheet\IOFactory;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
+use PhpOffice\PhpSpreadsheet\Writer\Csv;
 use PhpOffice\PhpSpreadsheet\Writer\IWriter;
 use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Contracts\Translation\TranslatorInterface;
+use PhpOffice\PhpSpreadsheet\Writer\Exception as PhpSpreadsheetException;
 
 class BillingService
 {
-    private const BATCH_SIZE = 200;
-    private const MAX_RESULTS = 50;
-
     public function __construct(
-        private readonly ApiServiceInterface $apiService,
-        private readonly ProjectRepository $projectRepository,
-        private readonly ClientRepository $clientRepository,
         private readonly InvoiceRepository $invoiceRepository,
         private readonly InvoiceEntryRepository $invoiceEntryRepository,
-        private readonly VersionRepository $versionRepository,
-        private readonly WorklogRepository $worklogRepository,
-        private readonly IssueRepository $issueRepository,
-        private readonly EntityManagerInterface $entityManager,
         private readonly TranslatorInterface $translator,
-        private readonly AccountRepository $accountRepository,
         private readonly string $invoiceSupplierAccount,
     ) {
     }
 
     /**
-     * Migrate from invoice.customer_account_id to invoice.client.
+     * Update total price for invoice entry by multiplying amount with price.
      */
-    public function migrateCustomers(): void
-    {
-        $invoices = $this->invoiceRepository->findAll();
-
-        /** @var Invoice $invoice */
-        foreach ($invoices as $invoice) {
-            $customerAccountId = $invoice->getCustomerAccountId();
-
-            if (null == $invoice->getClient() && null !== $customerAccountId) {
-                $client = $this->clientRepository->findOneBy(['projectTrackerId' => $invoice->getCustomerAccountId()]);
-
-                if (null !== $client) {
-                    $invoice->setClient($client);
-                }
-            }
-        }
-
-        $this->entityManager->flush();
-    }
-
-    /**
-     * @throws \Exception
-     */
-    public function syncIssuesForProject(int $projectId, callable $progressCallback = null): void
-    {
-        $project = $this->projectRepository->find($projectId);
-
-        if (!$project) {
-            throw new \Exception('Project not found');
-        }
-
-        $projectTrackerId = $project->getProjectTrackerId();
-
-        if (null === $projectTrackerId) {
-            throw new \Exception('ProjectTrackerId not set');
-        }
-
-        $issuesProcessed = 0;
-
-        $startAt = 0;
-
-        do {
-            $project = $this->projectRepository->find($projectId);
-
-            if (!$project) {
-                throw new \Exception('Project not found');
-            }
-
-            $pagedIssueData = $this->apiService->getIssuesDataForProjectPaged($projectTrackerId, $startAt, self::MAX_RESULTS);
-            $total = $pagedIssueData->total;
-
-            $issueData = $pagedIssueData->items;
-
-            foreach ($issueData as $issueDatum) {
-                $issue = $this->issueRepository->findOneBy(['projectTrackerId' => $issueDatum->projectTrackerId]);
-
-                if (!$issue) {
-                    $issue = new Issue();
-
-                    $this->entityManager->persist($issue);
-                }
-
-                $issue->setName($issueDatum->name);
-                $issue->setAccountId($issueDatum->accountId);
-                $issue->setAccountKey($issueDatum->accountKey);
-                $issue->setEpicKey($issueDatum->epicKey);
-                $issue->setEpicName($issueDatum->epicName);
-                $issue->setProject($project);
-                $issue->setProjectTrackerId($issueDatum->projectTrackerId);
-                $issue->setProjectTrackerKey($issueDatum->projectTrackerKey);
-                $issue->setResolutionDate($issueDatum->resolutionDate);
-                $issue->setStatus($issueDatum->status);
-
-                if (null == $issue->getSource()) {
-                    $issue->setSource($this->apiService->getProjectTrackerIdentifier());
-                }
-
-                foreach ($issueDatum->versions as $versionData) {
-                    $version = $this->versionRepository->findOneBy(['projectTrackerId' => $versionData->projectTrackerId]);
-
-                    if (null !== $version) {
-                        $issue->addVersion($version);
-                    }
-                }
-
-                if (null !== $progressCallback) {
-                    $progressCallback($issuesProcessed, $total);
-                    ++$issuesProcessed;
-                }
-            }
-
-            $startAt += self::MAX_RESULTS;
-
-            $this->entityManager->flush();
-            $this->entityManager->clear();
-        } while ($startAt < $total);
-
-        $this->entityManager->flush();
-        $this->entityManager->clear();
-    }
-
-    /**
-     * @throws \Exception
-     */
-    public function syncWorklogsForProject(int $projectId, callable $progressCallback = null): void
-    {
-        $project = $this->projectRepository->find($projectId);
-
-        if (!$project) {
-            throw new \Exception('Project not found');
-        }
-
-        $projectTrackerId = $project->getProjectTrackerId();
-
-        if (null === $projectTrackerId) {
-            throw new \Exception('ProjectTrackerId not set');
-        }
-
-        $worklogData = $this->apiService->getWorklogDataForProject($projectTrackerId);
-        $worklogsAdded = 0;
-
-        foreach ($worklogData as $worklogDatum) {
-            $project = $this->projectRepository->find($projectId);
-
-            if (!$project) {
-                throw new \Exception('Project not found');
-            }
-
-            $worklog = $this->worklogRepository->findOneBy(['worklogId' => $worklogDatum->projectTrackerId]);
-
-            if (!$worklog) {
-                $worklog = new Worklog();
-
-                $this->entityManager->persist($worklog);
-            }
-
-            $worklog->setWorklogId($worklogDatum->projectTrackerId);
-            $worklog->setDescription($worklogDatum->comment);
-            $worklog->setWorker($worklogDatum->worker);
-            $worklog->setStarted($worklogDatum->started);
-            $worklog->setProjectTrackerIssueId($worklogDatum->projectTrackerIssueId);
-            $worklog->setTimeSpentSeconds($worklogDatum->timeSpentSeconds);
-
-            if (null !== $worklog->getProjectTrackerIssueId()) {
-                $issue = $this->issueRepository->findOneBy(['projectTrackerId' => $worklog->getProjectTrackerIssueId()]);
-                $worklog->setIssue($issue);
-            }
-
-            if (!$worklog->isBilled() && $worklogDatum->projectTrackerIsBilled) {
-                $worklog->setIsBilled(true);
-                $worklog->setBilledSeconds($worklogDatum->timeSpentSeconds);
-            }
-
-            if (null == $worklog->getSource()) {
-                $worklog->setSource($this->apiService->getProjectTrackerIdentifier());
-            }
-
-            $project->addWorklog($worklog);
-
-            if (null !== $progressCallback) {
-                $progressCallback($worklogsAdded, count($worklogData));
-
-                ++$worklogsAdded;
-            }
-
-            // Flush and clear for each batch.
-            if (0 === $worklogsAdded % self::BATCH_SIZE) {
-                $this->entityManager->flush();
-                $this->entityManager->clear();
-            }
-        }
-
-        $this->entityManager->flush();
-        $this->entityManager->clear();
-    }
-
     public function updateInvoiceEntryTotalPrice(InvoiceEntry $invoiceEntry): void
     {
         if (InvoiceEntryTypeEnum::WORKLOG === $invoiceEntry->getEntryType()) {
@@ -248,6 +53,9 @@ class BillingService
         }
     }
 
+    /**
+     * Update total price for invoice by summing invoice entry total prices.
+     */
     public function updateInvoiceTotalPrice(Invoice $invoice): void
     {
         $totalPrice = 0;
@@ -259,112 +67,6 @@ class BillingService
         $invoice->setTotalPrice($totalPrice);
 
         $this->invoiceRepository->save($invoice, true);
-    }
-
-    public function syncAccounts(callable $progressCallback): void
-    {
-        $projectTrackerIdentifier = $this->apiService->getProjectTrackerIdentifier();
-
-        // Get all accounts from ApiService.
-        $allAccountData = $this->apiService->getAllAccountData();
-
-        foreach ($allAccountData as $index => $accountDatum) {
-            $account = $this->accountRepository->findOneBy(['projectTrackerId' => $accountDatum->projectTrackerId, 'source' => $projectTrackerIdentifier]);
-
-            if (!$account) {
-                $account = new Account();
-                $account->setSource($projectTrackerIdentifier);
-                $account->setProjectTrackerId($accountDatum->projectTrackerId);
-
-                $this->entityManager->persist($account);
-            }
-
-            $account->setName($accountDatum->name);
-            $account->setValue($accountDatum->value);
-            $account->setStatus($accountDatum->status);
-            $account->setCategory($accountDatum->category);
-
-            // Flush and clear for each batch.
-            if (0 === intval($index) % self::BATCH_SIZE) {
-                $this->entityManager->flush();
-                $this->entityManager->clear();
-            }
-
-            $progressCallback($index, count($allAccountData));
-        }
-
-        $this->entityManager->flush();
-        $this->entityManager->clear();
-    }
-
-    public function syncProjects(callable $progressCallback): void
-    {
-        // Get all projects from ApiService.
-        $allProjectData = $this->apiService->getAllProjectData();
-
-        foreach ($allProjectData as $index => $projectDatum) {
-            $project = $this->projectRepository->findOneBy(['projectTrackerId' => $projectDatum->projectTrackerId]);
-
-            if (!$project) {
-                $project = new Project();
-                $this->entityManager->persist($project);
-            }
-
-            $project->setName($projectDatum->name);
-            $project->setProjectTrackerId($projectDatum->projectTrackerId);
-            $project->setProjectTrackerKey($projectDatum->projectTrackerKey);
-            $project->setProjectTrackerProjectUrl($projectDatum->projectTrackerProjectUrl);
-
-            foreach ($projectDatum->versions as $versionData) {
-                $version = $this->versionRepository->findOneBy(['projectTrackerId' => $versionData->projectTrackerId]);
-
-                if (!$version) {
-                    $version = new Version();
-                    $this->entityManager->persist($version);
-                }
-
-                $version->setName($versionData->name);
-                $version->setProjectTrackerId($versionData->projectTrackerId);
-                $version->setProject($project);
-            }
-
-            $projectClientData = $this->apiService->getClientDataForProject($projectDatum->projectTrackerId);
-
-            foreach ($projectClientData as $clientData) {
-                $client = $this->clientRepository->findOneBy(['projectTrackerId' => $clientData->projectTrackerId]);
-
-                if (!$client) {
-                    $client = new Client();
-                    $client->setProjectTrackerId($clientData->projectTrackerId);
-                    $this->entityManager->persist($client);
-                }
-
-                $client->setName($clientData->name);
-                $client->setContact($clientData->contact);
-                $client->setAccount($clientData->account);
-                $client->setType($clientData->type);
-                $client->setPsp($clientData->psp);
-                $client->setEan($clientData->ean);
-                $client->setStandardPrice($clientData->standardPrice);
-                $client->setCustomerKey($clientData->customerKey);
-                $client->setSalesChannel($clientData->salesChannel);
-
-                if (!$client->getProjects()->contains($client)) {
-                    $client->addProject($project);
-                }
-            }
-
-            // Flush and clear for each batch.
-            if (0 === intval($index) % self::BATCH_SIZE) {
-                $this->entityManager->flush();
-                $this->entityManager->clear();
-            }
-
-            $progressCallback($index, count($allProjectData));
-        }
-
-        $this->entityManager->flush();
-        $this->entityManager->clear();
     }
 
     /**
@@ -441,15 +143,106 @@ class BillingService
     }
 
     /**
+     * Create a spreadsheet response from an array of invoice ids.
+     *
+     * @param array $ids Array of invoice ids.
+     *
+     * @throws EconomicsException
+     */
+    public function generateSpreadsheetCsvResponse(array $ids): Response
+    {
+        try {
+            $spreadsheet = $this->exportInvoicesToSpreadsheet($ids);
+
+            /** @var Csv $writer */
+            $writer = IOFactory::createWriter($spreadsheet, 'Csv');
+            $writer->setDelimiter(';');
+            $writer->setEnclosure('');
+            $writer->setLineEnding("\r\n");
+            $writer->setSheetIndex(0);
+
+            $csvOutput = $this->getSpreadsheetOutputAsString($writer);
+
+            // Change encoding to Windows-1252.
+            $csvOutputEncoded = mb_convert_encoding($csvOutput, 'Windows-1252');
+
+            $response = new Response($csvOutputEncoded);
+            $filename = 'invoices-'.date('d-m-Y').'.csv';
+
+            $response->headers->set('Content-Type', 'text/csv');
+            $response->headers->set('Content-Disposition', 'attachment; filename="'.$filename.'"');
+            $response->headers->set('Cache-Control', 'max-age=0');
+
+            return $response;
+        } catch (PhpSpreadsheetException) {
+            throw new EconomicsException(
+                $this->translator->trans('exception.invoices_export_failure'),
+                400
+            );
+        }
+    }
+
+    /**
+     * Create spreadsheet html from an array of invoice ids.
+     *
+     * @throws EconomicsException
+     */
+    public function generateSpreadsheetHtml(array $ids): bool|string
+    {
+        try {
+            $spreadsheet = $this->exportInvoicesToSpreadsheet($ids);
+
+            $writer = IOFactory::createWriter($spreadsheet, 'Html');
+
+            $html = $this->getSpreadsheetOutputAsString($writer);
+
+            if (empty($html)) {
+                $html = '<html lang="da" />';
+            }
+
+            // Extract body content.
+            $d = new \DOMDocument();
+            $mock = new \DOMDocument();
+            $d->loadHTML($html);
+            /** @var \DOMNode $body */
+            $body = $d->getElementsByTagName('div')->item(0);
+
+            foreach ($body->childNodes as $child) {
+                if ($child instanceof \DOMElement) {
+                    if ('table' == $child->tagName) {
+                        $child->setAttribute('class', 'table table-export');
+                    }
+                }
+                $mock->appendChild($mock->importNode($child, true));
+            }
+
+            $html = $mock->saveHTML();
+
+            if (!$html) {
+                throw new EconomicsException(
+                    $this->translator->trans('exception.invoices_export_failure_could_not_generate_html'),
+                    400
+                );
+            }
+
+            return $html;
+        } catch (PhpSpreadsheetException) {
+            throw new EconomicsException(
+                $this->translator->trans('exception.invoices_export_failure'),
+                400
+            );
+        }
+    }
+
+    /**
      * Export the selected invoices (by id) to csv.
      *
      * @param array $invoiceIds array of invoice ids that should be exported
      *
      * @return Spreadsheet
-     *
-     * @throws \Exception
+     * @throws EconomicsException
      */
-    public function exportInvoicesToSpreadsheet(array $invoiceIds): Spreadsheet
+    private function exportInvoicesToSpreadsheet(array $invoiceIds): Spreadsheet
     {
         $spreadsheet = new Spreadsheet();
         $sheet = $spreadsheet->getActiveSheet();
@@ -474,7 +267,7 @@ class BillingService
                 $client = $invoice->getClient();
 
                 if (is_null($client)) {
-                    throw new \Exception('Client cannot be null.');
+                    throw new EconomicsException('Client cannot be null.', 400);
                 }
 
                 $internal = ClientTypeEnum::INTERNAL === $client->getType();
@@ -587,9 +380,9 @@ class BillingService
     }
 
     /**
-     * @throws \PhpOffice\PhpSpreadsheet\Writer\Exception
+     * @throws PhpSpreadsheetException
      */
-    public function getSpreadsheetOutputAsString(IWriter $writer): string
+    private function getSpreadsheetOutputAsString(IWriter $writer): string
     {
         $filesystem = new Filesystem();
         $tempFilename = $filesystem->tempnam(sys_get_temp_dir(), 'export_');
